requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
backports.zoneinfo==0.2.1
Django==3.1.14
django-cors-headers==3.11.0
django-environ==0.4.5
django-graphql-jwt==0.3.1
djangorestframework==3.13.1
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
Pillow==9.0.0
promise==2.3
PyJWT==1.7.0
pytz==2022.6
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.2
text-unidecode==1.3
